==> 01-format-mermaid-public.R <==
# ---01-format-mermaid-public.R ---
#  Continue from 00-load-mermaid-public.R with summary *.csv files
#  W. Friedman // Sept 2021
# ---------------------------------

library(here)
library(tidyverse)
library(janitor)

# Load all files generated by 00-load-mermaid-public.R

pit <- read_csv(here("globalprep", "analysis", "coral-counterfactual", "all-public-summary.csv"))
lit <- read_csv(here("globalprep", "analysis", "coral-counterfactual", "esd-lit.csv"))
bleaching <- read_csv(here("globalprep", "analysis", "coral-counterfactual", "esd-bleaching.csv"))

# Explore

# View(pit)
# View(lit)
# View(bleaching)

# ~ 2000 clean site-year replicates
nrow(pit) # 1730
nrow(lit) # 27
nrow(bleaching) # 301

# Years & Countries
bleaching %>% 
  mutate(year = lubridate::year(sample_date)) %>% 
  filter(!is.na(percent_hard_avg_avg)) %>%  #279 rows
  tabyl(country, year) %>% 
  adorn_totals()

pit %>% 
  mutate(year = lubridate::year(sample_date)) %>% 
  filter(!is.na(percent_cover_benthic_category_avg_hard_coral)) %>%  #279 rows
  tabyl(country,year) %>% 
  adorn_totals()

lit %>% 
  mutate(year = lubridate::year(sample_date)) %>% 
  filter(!is.na(percent_cover_benthic_category_avg_hard_coral)) %>%  #279 rows
  tabyl(country,year) %>% 
  adorn_totals()


# Check columns
names(pit) %in% names(lit) # 1 column is different
names(pit)[!names(pit) %in% names(lit)]  #data_policy_benthicpit
names(lit)[!names(lit) %in% names(pit)]  #data_policy_benthiclit

names(pit) %in% names(bleaching) # several are different
names(pit)[!names(pit) %in% names(bleaching)]
names(bleaching)[!names(bleaching) %in% names(pit)]

## From Emily: NB this is comparable to percent_cover_benthic_category_avg_hard_coral
## (WF: presumably similar for macroalgae, soft_coral,)
bleaching$percent_hard_avg_avg 


# Combine into one database of coral cover
# Columns to keep (all for now)?
# Create new variable 'pct_cover_hard_coral_combined' with values from all 3 datasets.

cc_mermaid <- pit %>%
  mutate(db  = "pit",
         pct_cover_hard_coral_combined = percent_cover_benthic_category_avg_hard_coral) %>%
  full_join(lit %>% 
              mutate(db =  "lit",
                     pct_cover_hard_coral_combined = percent_cover_benthic_category_avg_hard_coral)) %>% 
  full_join(bleaching %>% 
              mutate(db =  "bleaching",
                     pct_cover_hard_coral_combined = percent_hard_avg_avg))

cc_mermaid %>% tabyl(db)
cc_mermaid %>% tabyl(pct_cover_hard_coral_combined, db)

rm(pit,lit,bleaching)


# KEEP ONLY surveys where hard coral is recorded.
# Create 'year' column.
cc_mermaid <- cc_mermaid %>% 
  mutate(year = lubridate::year(sample_date)) %>% 
  drop_na(pct_cover_hard_coral_combined)

cc_mermaid

cc_mermaid %>% tabyl(db, year)

# Add new transect id ('id_new'); see notes on dupes below.
cc_mermaid <- cc_mermaid %>% 
  arrange(country, site, year, sample_date, db) %>% 
  mutate(cc_id = seq(1:nrow(cc_mermaid)), 
         cc_id = str_c("mermaid_",cc_id))

cc_mermaid %>% 
  write_csv(here("globalprep", "analysis", "coral-counterfactual", "outputs", 
                 "combined MERMAID cc_data.csv"))

cc_mermaid %>% 
  tabyl(country)
